NullReferenceException Fix
By surrounding the unsafe code with try block, you can try to execute the unsafe code, and if it fails, then the regular method will be used.

I had some problems with unlocking bits as well, you should definitely look into that. I have surrounded that part of the code with a try block because it occasionally threw the System.Runtime.InteropServices.ExternalException